Native Spirng Security Login is working using user that were direcly injected into MySQL..Used generatePassword Unit test to get the password hash to manually enter. Lots still broken, like custom login, and registration
