+++ src/main/java/com/liftoff/controllers/HelpController.java
@@ -1,10 +1,11 @@
 package com.liftoff.controllers;
 
-import com.liftoff.models.Welcome;
+
 import com.liftoff.models.data.FaqRepository;
 import com.liftoff.models.Faq;
 import com.liftoff.models.data.WelcomeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,6 +23,8 @@ public class HelpController {
 
     @Autowired
     private WelcomeRepository welcomeRepository;
+
+
 
     @GetMapping ("")
     public String displayHelpPage (Model model) {
