+++ src/main/java/com/liftoff/models/data/RoleRepository.java
@@ -1,0 +1,7 @@
+package com.liftoff.models.data;
+
+import com.liftoff.models.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RoleRepository extends JpaRepository<Role, Integer> {
+}
