+++ src/main/java/com/liftoff/security/WebSecurityConfig.java
@@ -58,6 +58,13 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
                     .antMatchers("/","/about","/help/**","/wantTo/**","/css/*","/js/*","/register").permitAll()
                     .anyRequest()
                     .authenticated()
+//                .and()
+//                .formLogin()
+//                    .loginPage("/login")
+//                    .permitAll()
+//                    .defaultSuccessUrl("/access",true)
+//                    .passwordParameter("password")
+//                    .usernameParameter("username")
                 .and()
                 .formLogin().permitAll()
                 .and()
