+++ src/main/resources/templates/users/index.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org/"
+      xmlns:sec="http:/www.thymeleaf.org/thymeleaf-extras-springsecurity5">
+    <head th:replace="fragments :: head"></head>
+
+    <body class="container text-center">
+
+        <header th:replace ="fragments :: header"></header>
+
+        <table class="table table-bordered">
+            <thead>
+                <tr>
+                    <th>ID</th>
+                    <th>username</th>
+                    <th>email</th>
+                    <th>hashed password</th>
+                    <th>Roles</th>
+                    <th sec:auth ></th>
+
+                </tr>
+            </thead>
+            <tbody>
+                <th:block th:each="user : ${listUsers}">
+                    <tr>
+                        <td th:text="${user.id}"></td>
+                        <td>[[${user.username}]]</td>
+                        <td>[[${user.email}]]</td>
+                        <td>[[${user.pwHash}]]</td>
+                        <td>[[${user.roles}]]</td>
+
+                        <td >
+                            <a th:href="@{'/users/edit/' + ${user.id}}">edit</a>
+                            &nbsp;
+                            <a th:href="@{'/users/delete/' + ${user.id}}">delete</a>
+                        </td>
+
+                    </tr>
+                </th:block>
+            </tbody>
+        </table>
+    </body>
+</html>
